Jashua - 7 al 15 de Enero
Jan 15 commit 83d47e9fa7b8f97b85955acc2cae6719f0214bf9

Manager Sales

Agregado el JS de manager_sales al header
Agregada la funcion de borrado en ventas

Jan 14 commit 6e5ef81b84578584036c88e32988e51e4aa36734

Paypal

Ahora guarda la venta de paypal en la tabla de ventas normales

Jan 14 commit b3c75eb468fc506fedb1d44ba02745f9ddf77444

Paypal

Agregado el modelo que regresa la info del error de paypal
Agregada la funcion que arregla los errores de paypal
Removidas las ejecuciones de la funcion consoleLog

Jan 14 commit 47ab381ea62fedc1c8ad03523c9e3fd640e1840e

Paypal

Removida la opcion de pagar sin cuenta

Jan 14 commit 6c0b62d4461be9c5c1dfd60de299bf919336a277

Manager Sales

Creadas las view, modelo y controlador
Agregado el campo de paypal id en ventas (bd)

Jan 14 commit 209dbbefb788a2d16920d6111de3ce1fbf799e28

Manager Clients

Terminadas las funciones basicas de la tabla

Jan 13 commit d53b8acfe0fdc23ab9e742d2749c5124fd66a265

Manager Clients

Creadas las vistas
Creado el modelo
Creado el archivo js

Jan 13 commit cb9b01ce564558049fdec16e51e84b32362dda2c

Manager Users

Validacion basica en agregar y editar usuarios

Jan 13 commit 0018a92b1a3db3a1b9c5c20aa2a6acea0bc1ce8c

Managers Users

Borra al usuario correctamente

Jan 13 commit d9193b799c60ec778f78cb95fd5ad3fdc2d8e9f9

Manager Users

Edita al usuario correctamente (sin validar)
Username fue cambiado a Email

Bug: No muestra la cpassword
Causa: Falta encriptar en el registro

Jan 10 commit 625c800a801d16cdcc9b3120aedc51145c7fac07

Manager Users

Agrega correctamente los usuarios (sin validar)
Removido rol de usuario al agregar desde manager
Agregado manager_users.js al footer de manager
Actualizados los enlaces al controlador Manager_Users

Jan 10 commit 6cb230ebdf9cb243af40fd92da77707bf8b09ce0

Manager Usuarios

Modificado el modelo de manager usuarios
Modificado el controlador de manager usuarios

Jan 10 commit 1e4300581ec1d3b39a97f920babdfa51077f4976

Manager Usuarios

Views de manager usuarios separados a otra carpeta
El header apunta al nuevo controlador
La datatable muestra usuarios (sin funcionalidad)
Agregado el campo de telefno en los view

Jan 10 commit 8730c60213cf2e1a49af16451a9398c5f74e1fdf

Manager Users

Separado los views del manager de usuarios (no borre views)
Separadas las funciones del modelo manager para usuarios
Refactorizado el controlador de manager para usuarios
Falta refactorizar algunos modelos

Jan 9 commit e4cc45274d629adb01753ff756a705a818cc360d

Manager

Separado el controlador, ahora hay un manager para usuarios
Agregado el helper para la activacion del header
Refactorizando codigo en progreso

Jan 9 commit a54a485766060e1ec06254303e44c96852d5ce31

Registro

El email enviado para validacion ahora es UTF-8
Sesion username cambiada por el email del usuario
Comentarios agregados al controlador de Login

Jan 9 commit 32447ee9a189e3dcac5a15ecd17af9580a804c94

Registro usuario

Envia el email correctamente con la informacion en el enlace
Actualiza la verificacion en la base de datos

Los siguientes casos estan cubiertos:
    1) Email no existe
    2) Usuario ya esta verificado
    3) Usuario verificado
    4) Error de servidor
    X) Error de informacion

Jan 7 commit 18cabe832b214d230b03525050a80095c6874708

Modificados los PK de YOY Paypal

Jan 7 commit 307b5bd67bfb8f63c1adb2f3369a8bf36df1aa4c

Email y SQL

Ya manda los correos de manera correcta
Campo de codigo agregado a la bd

Jan 7 commit fe48d442019043175bc1f4d4a2468400209ca34c

Registro

Ahora verifica que el email no exista en la base de datos
antes de ser registrado.
Se registran con vigencia de 1

Jan 7 commit a5ec030f58bf8c6c23b9480e298ebd5a0937ce13

Login y Registro

Inicio de sesion con password encriptada
Telefono ya no es requerido

+++ other/YOY_paypaltables.sql
@@ -8,7 +8,7 @@ SET FOREIGN_KEY_CHECKS=0;
 DROP TABLE IF EXISTS 	`paypal_client`;
 CREATE TABLE 			`paypal_client` (
 
-	`paypal_client_id` 	TINYINT PRIMARY KEY NOT NULL AUTO_INCREMENT,
+	`paypal_client_id` 	INT PRIMARY KEY NOT NULL AUTO_INCREMENT,
     
     `id` 				VARCHAR(25) UNIQUE NOT NULL,
     `name` 				VARCHAR(50) NOT NULL,
@@ -27,9 +27,9 @@ CREATE TABLE 			`paypal_client` (
 DROP TABLE IF EXISTS 	`paypal_order`;
 CREATE TABLE 			`paypal_order` (
 
-	`paypal_order_id` 	TINYINT PRIMARY KEY NOT NULL AUTO_INCREMENT,
+	`paypal_order_id` 	INT PRIMARY KEY NOT NULL AUTO_INCREMENT,
     
-    `paypal_client_id` 	TINYINT DEFAULT NULL,
+    `paypal_client_id` 	INT DEFAULT NULL,
     `ID_USUARIO`		INT DEFAULT NULL,
     `ID_PRODUCTO` 		INT DEFAULT NULL,
     
@@ -66,7 +66,7 @@ CREATE TABLE 			`paypal_order` (
 DROP TABLE IF EXISTS 	`paypal_error`;
 CREATE TABLE 			`paypal_error` (
 
-	`paypal_error_id` 	TINYINT PRIMARY KEY NOT NULL AUTO_INCREMENT,
+	`paypal_error_id` 	INT PRIMARY KEY NOT NULL AUTO_INCREMENT,
 	
 	`ID_USUARIO`		INT NOT NULL,
     `ID_PRODUCTO` 		INT NOT NULL,
